requirements.txt
aiogram>=3.4.0
aiosqlite>=0.19.0
apscheduler>=3.10.0
aiohttp>=3.9.0
gTTS>=2.5.0
Pillow>=10.2.0
python-dotenv>=1.0.0
SQLAlchemy>=2.0.0
alembic>=1.13.0
asyncpg>=0.29.0